Move UserData into types fodler
