+++ app/src/types/UserData.ts
@@ -1,0 +1,12 @@
+interface RoleData {
+    name: string;
+    description: string;
+  }
+  
+  export interface UserData {
+    id: string;
+    username: string;
+    display_name: string;
+    is_active: boolean;
+    roles: RoleData[];
+  }
